create navbar component and replace issues with nav component
