+++ src/pages/issue-1.js
@@ -1,18 +1,11 @@
 import React, { useEffect, useState } from "react"
-import { useMediaQuery } from "react-responsive"
 
 import SEO from "../components/seo"
-import { NavBar, StyledLink } from "../styles/navigation"
-import { Text } from "../styles/text"
-import { breakpoints } from "../styles/layout"
 import Issue1 from "../issues/issue-1"
-import logo from "../images/logo.svg"
+import Nav from "./navBar"
 
 const PublishedIssue1 = () => {
   const [isLoading, setIsLoading] = useState(true)
-  const isPhoneWide = useMediaQuery({
-    query: breakpoints.phoneWide,
-  })
 
   useEffect(() => {
     setIsLoading(false)
@@ -25,15 +18,7 @@ const PublishedIssue1 = () => {
   return (
     <>
       <SEO title="Home" />
-      <NavBar>
-        <img
-          width={isPhoneWide ? 75 : 128}
-          src={logo}
-          alt="small qrzn logo"
-        ></img>
-        {!isPhoneWide && <Text center>the quaranzine for the quaranzined</Text>}
-        <StyledLink to="/archive/">archive</StyledLink>
-      </NavBar>
+      <Nav />
 
       <Issue1 />
     </>
