+++ src/pages/issue-2.js
@@ -1,19 +1,11 @@
 import React, { useEffect, useState } from "react"
-import { useMediaQuery } from "react-responsive"
 
 import SEO from "../components/seo"
-import { NavBar, StyledLink } from "../styles/navigation"
-import { Text } from "../styles/text"
-import { breakpoints } from "../styles/layout"
-import logo from "../images/logo.svg"
-
+import Nav from "./navBar"
 import Issue2 from "../issues/issue-2"
 
 const PublishedIssue2 = () => {
   const [isLoading, setIsLoading] = useState(true)
-  const isPhoneWide = useMediaQuery({
-    query: breakpoints.phoneWide,
-  })
 
   useEffect(() => {
     setIsLoading(false)
@@ -26,15 +18,7 @@ const PublishedIssue2 = () => {
   return (
     <>
       <SEO title="Home" />
-      <NavBar>
-        <img
-          width={isPhoneWide ? 75 : 128}
-          src={logo}
-          alt="small qrzn logo"
-        ></img>
-        {!isPhoneWide && <Text center>the quaranzine for the quaranzined</Text>}
-        <StyledLink to="/archive/">archive</StyledLink>
-      </NavBar>
+      <Nav />
       <Issue2 />
     </>
   )
